Optimize query for application reviewers to include count of owned pending applications
The staff dashboard needs to display the number of pending applications
owned by a particular application reviewer. That lets us know how many
applications are assigned to us at a glance.

Optimize the database query that searches for the application reviewer
objects for this round. Make the database query annotate the query with
the number of pending applications that reviewer owns. That way we look
up all the information we need in one query, rather than querying for
each reviewer.