Remove LibrePlanet listing from Outreachy travel stipend page.
In December 2018, former LibrePlanet speakers called on LibrePlanet to
bar Richard Stallman from attending LibrePlanet:

https://wwahammy.com/on-safety-at-libreplanet/

Stallman said in talks that as President of the FSF, he was not subject
to the Code of Conduct:

https://media.libreplanet.org/u/libreplanet/m/the-gnu-philosophy-ethics-beyond-ethics/

As of March 23, 2021, Stallman has been reinstated to the FSF board:

https://web.archive.org/web/20210322224339/https://www.fsf.org/about/staff-and-board/

Based on past behavior, Outreachy is concerned Stallman will use his
position as a reinstated FSF board member to continue to say he is not
subject to LibrePlanet Code of Conduct.

Therefore, Outreachy cannot recommend LibrePlanet to Outreachy interns
and alums.