Update initial feedback forms and hours interns report working
Simplify the initial feedback form. Ask the following new questions:

 * Do mentor(s) answer the intern's questions within 10 hours?
 * Does the intern ask mentor(s) questions when stuck for more than 1 to 3 hours?
 * Do mentor(s) offer more support if the intern is stuck?
 * Do mentor(s) meet privately with the intern?
 * Do mentor(s) meet with the intern over phone or video chat?
 * Has the intern recently missed more than 2 meetings?
 * Do mentor(s) and the intern talk about project progress at least 3 days a week?
 * Has the intern created a blog?

Allow interns to report working 45 or 55 hours per week. In the past we
asked if they worked 5, 10, 15 ... 35, 40, 50, 60 hours per week. Add
the missing two values in the range.

Modify the staff view of the feedback to highlight when an intern is
working overtime. In the past, we didn't set a warning color if the
intern reported working overtime during the midpoint or final feedback.

Update the initial feedback instructions to reflect the new questions.

Update the admin dashboard to show the new initial feedback objects
inline in the InternSelection objects.

Delete the form templates for the old initial feedback format.

I reviewed the privacy policy, and it generically references feedback
data gathered, rather than listing the specific questions we ask.
Therefore, no changes are needed to the privacy policy.