Show total number of interns approved on community sponsor summary page
Sometimes a community finds sponsorship for a number of interns, but
then accepts less interns than they expected. If a sponsor has committed
to funding a specific community, in this case we only invoice the
sponsor for the number of interns the community accepted.

Add the total number of interns accepted and the total amount of funding
needed for that number of interns to the sponsorship summary page. This
will allow organizers to visually see whether sponsors should be
invoiced for less than the amount of sponsorship they said they could
sponsor.