Add sponsorship details page with revision history.
Outreachy organizers send sponsorship invoice requests to our fiscal
non-profit parent organization, Software Freedom Conservancy.

When Outreachy organizers send invoice requests, we need to provide
proof that sponsors have committed to funding Outreachy. This is part of
Conservancy's bookkeeping process. As a non-profit, they need to provide
tax auditors proof that Conservancy was invoicing donors at their request.

That means Outreachy needs to provide proof that donors asked to sponsor
Outreachy interns for a particular community.

In the easiest case, the sponsor emails the Outreachy organizers
expressing interest in sponsoring. Then we forward the email to
Conservancy when we ask them to send an invoice to the sponsor.

It gets more complicated to provide proof of sponsorship when community
coordinators enter sponsor information on the Outreachy website. When a
community signs up to participate as a mentoring organization, the
website asks them to provide sponsorship information (sponsor name,
amount of sponsorship, and any additional details).

The Outreachy website records the revision history of the sponsorship
object in the Outreachy database. We track which website account edited
that object, the date and time it was edited, and what data the object
contained after editing.

We can use this revision history to prove that community coordinators
asserted to Outreachy organizers that a specific sponsor would fund
Outreachy interns with their community.

Create a new page per community per round that will display sponsorship
details, including the sponsorship revision history. That will allow
Outreachy organizers to print out the page as proof of sponsorship for
the invoice request to Conservancy.