Add fields in InternSelection class to track payment status.
In the past, we used a nullable BooleanField to represent payment
status:
 - "None" meant the mentor feedback was under review by organizers
 - "True" meant organizers had requested the intern be paid the stipend
 - "False" mean organizers had requested the intern not be paid

Now we want to make mentor feedback review a multi-step process, so we
need more feedback/payment status states. Therefore, we need to migrate
from a nullable BooleanField to a CharField choice field.

Make a CharField for the initial, midpoint, and final payment status:
 - WAITINGONFEEDBACK - Waiting for new or updated feedback from the mentor
 - UNDERREVIEW - Feedback about intern has not been reviewed
 - QUEUED - Feedback reviewed, payment request queued to be sent to Conservancy
 - REQUESTED - Payment request sent to Conservancy
 - DONOTPAY - Conservancy has been notified not to pay the intern stipend

While we're at it, move the payment and feedback status from the
Initial/Midpoint/FinalMentorFeedback classes into the InternSelection
class. This will allow us to track status without having to catch an
exception if the feedback object doesn't exist. It also means we may be
able to simplify the code that finds all feedback version that were
edited by mentors (vs. organizers setting organizer_payment_approved
through the Django admin dashboard).

Make a custom migration to set the new payment status fields based on
the status stored in the newly deprecated organizer_payment_approved
field.

Once I've checked this code correctly sets the status fields on a test
database, I'll push a second commit that removes the
organizer_payment_approved field. That way we don't lose data.