Change intern and mentor feedback "buttons" that do nothing to badges
When writing the templates to display intern and mentor feedback status,
I used HTML buttons instead of span badge elements. This caused it to
look like you could "click" on the feedback status.

Using buttons instead of badges was very confusing when the status was
"pay". It made it look like you could click the button to send an email
to authorize payment. Instead, the button was supposed to indicate that
the mentor feedback was to pay the intern.

Change to use badges to indicate the feedback status.