Rewrite community sign up page.
Simplify the community sign up page. Move wordy sections into the
community documentation and link them.

Add more rules and common misconceptions to the page. Hopefully this
will discourage people with ineligible communities that want Outreachy
general funding not to sign up.

Add a section at the top of the page and the top of the community
documentation to define what "communities" mean. We use the word so
much, but people who aren't part of open source don't understand what it
means. They think that a community is a city or a local community
organization. Make it clear that a "community" in Outreachy is either an
open source community or a non-profit.

Add documentation about Outreachy interns not being contractors for
communities or paid employees for sponsors. That should discourage
communities that want to "hire" only experienced interns and can't
provide feedback to improve that intern's skills. Outreachy isn't a
work-for-hire arrangement.

Add a note that Outreachy may consider non-American charitable
organizations to participate.