Ask new communities to provide repo links and lists of licenses.
One of the most tedious parts of new community review is actually trying
to find their source code (or creative works repo). Ask new communities
to provide links to their repositories. Ask them to provide a list of
the open source licenses used in those repos.

There are no tests to update with these fields because we don't have
tests for adding a new community. 😅 I tested this locally, but we
really should add some tests for this. I think the reason they're hard
to test is because there are several forms to test.