Revamp initial application.
Remove questions about how applicants contributed to open source. We
haven't used that data since we've been collecting it. Keep general
questions about whether they contributed to open source, were paid to
contribute to open source, and if they participated in GSoC or
Outreachy.

Remove help text for a lot of the questions. We often need to change the
wording in the middle of the application period if we find it was
confusing. This means we won't have to do a database migration if we
change the help text.

Add a field for organizers to record notes about official universities.
This will allow us to tell initial application reviewers information
about student eligibility when some students are eligible but others
aren't.

Add a field to the essay questions section. Prompt applicants for
content warnings about their essay answers. Try to word it so they don't
hesitate to submit their essay. Don't use the word "traumatic", as some
people may not think of their experiences as traumatic.

Clarify the help text around time commitments. Try to clear up common
questions, like "If I will quit my job, should I say I'm an employee?"

Clarify instructions for volunteer questions.

Make it more clear when you are on the last page of the initial
application.

Add instructions to the first page for which sections of the Applicant
Guide to read.

The instructions for the school term dates and the contractor time
commitments remain an absolute disaster. Someone please simplify the
English on these pages.

I will sleep now, and hope I haven't messed things up.