Week 13 email: Remove travel stipend reference, other edits.
Remove references to the travel stipend. We stopped the travel stipend
because of the pandemic, and we have decided to not reinstate it. This
means interns get paid more, even if they don't travel to conferences.
It also means less overhead for Conservancy accounting to deal with
travel reimbursement requests.

Add chat topic: staying involved with Outreachy.

s/free software/open source/