Show payment schedule and feedback for 60 days after internship ends
On the intern and mentor dashboard, show the internship feedback status
and payment schedule for 60 days after the final feedback is due. This
*should* be long enough for any intern to be paid the final internship
stipend, even taking into account corner cases, like bank errors or
Outreachy organizers not processing feedback promptly.

This does mean mentors who participate in back-to-back cohorts may see
the payment details and schedule for longer, which is a bit of an
annoyance. But I would rather show the information to mentors who are
trying to help interns who have not received their final payment.