Update Python runtime to 3.9.7
The latest version of herokuish (herokuish:v0.5.30-20) no longer
supports Python 3.7.3. So we need to update the Python version we're
using on the production server.

Debian testing currently has Python 3.9.7 available. Python 3.9.7 is the
latest Python version heroku supports, according to this page (as of
September 6, 2021):

https://devcenter.heroku.com/articles/python-support

Update the Heroku Python buildpack we fetch so that the scripts can find
the latest version of Python to install.