Make application review actions take a next URL
In upcoming changes, there will be additional views for processing the
initial application. Those templates still have links to rate the
application, change application owner, and update reviewer comments. We
want to redirect to the new URLs for application processing, instead of
the application preview URL.

Therefore, change the application review action views to redirect to a
next URL, if it is present. Otherwise, redirect to the application
preview URL.