Second try at fixing counts of applications to process.
When Outreachy organizers are processing initial applications, they only
want to see the applications that are not assigned to a reviewer.
Assigning it to a reviewer typically means they need a second review.

Clarify the dashboard heading that initial applications to process does
not include applications assigned to a reviewer.

Change the query for counting strong, good, maybe, and unclear initial
applications to filter for applications that are not assigned to a
reviewer.