Intern agreement: Update extension procedures to reflect current policies
In the past, internship extensions had to be approved by the Outreachy
Project Leadership committee (PLC). Now Outreachy organizers review the
internship extensions first.

If Outreachy organizers believe the extension should be approved, they
will notify the mentor and intern that the extension is granted.

If they are unsure of whether an internship should be extended or they
believe the internship extension should be denied, they will determine
the action to take by discussing it with the Outreachy PLC.