Interns will now be paid a $7,000 USD stipend
Update language on the website to reflect the change from a $6,000 USD
stipend to a $7,000 USD stipend.

The mentor and sponsor page breaks down the sponsorship amount of $6,500
into the internship stipend and a $500 admin fee. That text no longer
makes sense, given the increase in internship stipend total.

We will be working with communities and sponsors over the next year to
determine a new sponsorship model for the December 2022 cohort. In the
meantime, communities and sponsors will still sponsor at $6,500 per
intern.