Change internship effort from 40 hours/week to 30 hours/week
Change all text in the HTML templates to reflect our new 30 hours/week
internship effort requirement.

Some of the models are still incorrect, referencing 40 hours per week.
Most of that text we override in the template (e.g. in the initial
application first page that asks general eligibility questions). Since
that requires a migration, I'm hesitant to edit them so close to the
announcement date.

I did not look closely at the model code that calculates whether a
person is eligible due to their time commitments. That will need to be
revamped in December or January.