Fix payment and feedback schedule in internship guide
Outreachy will now collect feedback from mentors and interns 4 times
during the internship.

The initial stipend ($3,000) will be sent after feedback #1 is
successful.

Feedback #2 is a check point to ensure the internship is running
smoothly, but is not associated with a stipend payment.

The final stipend ($4,000) will be sent after feedback #3 is
successful.

Feedback #4 is a check point to ensure the internship is running
smoothly, but is not associated with a stipend payment.

Internships can be extended at any of the four feedback points.