Week 1 chat: change topic to remote work tips
Most Outreachy interns introduce themselves as soon as they sign into
the Zulip chat. They also tend to ask questions in the #help channel
right away. We've found that the intro Q&A chat tends to be a little
light on topics.

At the same time, we have also found that most interns and some mentors
are new to remote internships. So this cohort we want our first chat to
talk about how to effectively work remotely.

Remove mention of the chat topic from the first day email, so that if we
change the topic later, it's easier to edit it in one place.