Week 3 chat: change topic back to everyone struggles
For the May 2021 cohort, we tried changing the week 3 chat topic to be
about confusing open source vocabulary. While the chat was lively, we
noticed that we had a higher than normal amount of interns who were
struggling and not talking to their mentor, or who were shy about asking
questions of their mentor.

We think the week 3 chat really normalizes that it's okay to struggle,
and it's important that interns reach out for help when they're stuck.

We have received feedback that interns are uncomfortable sharing how
they're struggling if their mentors are participating in the chat. So we
are moving the chat into the interns & alums private channel.

There might be some mentors who are past Outreachy interns in that
channel, but I don't want to create a new channel every cohort for just
this chat. It's also a pain to add people to a stream outside of when
they're initial subscribed. You have to add people individually.

Interns also told us in the past that they were uncomfortable writing
about what they were struggling with. So we'll keep the blog post about
confusing open source vocabulary instead. I still shared the rubber duck
debugging article, because I like it a lot, but it may be less relevant.