Show intern blog link (or missing) in staff view of feedback #1
If the mentor and intern disagree about whether the intern has created a
blog, it can be helpful to check if the intern has recorded their blog
URL in their Outreachy website account. It can also be helpful to be
able to follow the link to the blog to see if the intern has created a
blog but not actually created blog posts yet.