Add emojis to community requirements, tweak eligibility rules emphasized
Remove the talk of communities required to not be tied to one company.
We received feedback that many open source communities in the global
south work this way, because they don't have the privilege to work on
open source for free. We will evaluate this rule on a case-by-case
basis.

The "common mistakes" section at the end should deter western companies
who are seeking employees from marginalized groups in tech.