More updates to the sponsorship page
Remove outdated references in the sponsor page to resources still being
on GNOME servers. While the IRC channel is still on irc.gnome.org, we
don't use it any more.

Remove reference to community credits in sponsorship page. We're trying
to get sponsors to fund the general fund, and doesn't make sense to
mention it here.

TODO: Add information about community credits to the Community Guide.

Mention we can invoice sponsors earlier or later to fit budget
deadlines.

Bold questions in FAQ to make it easy to scan.