Fix feedback #3 export
The Conservancy script to parse the Outreachy intern feedback json file
relies on the file name to determine which payment the feedback is for.
It associates the 'midpoint' feedback with payment #2.

For Outreachy, payment #2 is associated with feedback #3. Therefore, we
need to include "midpoint-feedback" in the json file name for feedback #3.

Legacy integration code is hard!