Store country and code from initial application
Outreachy's fiscal parent non-profit, Software Freedom Conservancy,
complies with laws that state U.S. based non-profits must not send money
to regions under U.S. sanctions.

In the past, when the list of sanctioned regions was (relatively)
stable, we asked people on their initial application 'Are you a citizen,
resident, or national of Crimea, Cuba, Iran, North Korea, or Syria'.
This would give us an indication of whether the person might be under
U.S. sanctions.

With recent world events, the list of countries and regions under U.S.
sanctions is rapidly changing. Unfortunately the list began changing
after initial applications had already been submitted. We expect it to
continue changing as the situation evolves.

Our best guess whether applicants live in the new sanctioned regions on
the list now comes from the country information they provided in the
initial application.

Originally, we asked for that information because it would often clarify
essays for application reviewers. We would delete the information about
which country the applicant will be living in during the internship when
we purged sensitive initial application information (including the
essays).

Now, we need to use that country information to help Conservancy comply
with U.S. sanctions laws. Update our privacy policy to clarify why we
collect this information and how it will be used.

Add a manual database migration that saves the country the applicant
will be living in during the internship in new ApplicantApproval fields.
If the essays have already been purged, this information will not be
saved.

Update the view code that runs when an applicant fills out an initial
application to save the country information in ApplicantApproval. Add a
test to ensure the information is saved, even after the sensitive
initial application data is purged.