Improve initial application results page.
Fix page to correctly show minimum free time required. A change to store
the application in the user's role object broke this page.

List approximate dates for future cohorts in both the alignment and time
rejected sections.

Improve the wording of the alignment rejected sections to be more
gentle. Try to make it easier to read for non-native English speakers.

Point people with alignment-rejected essays to the Applicant Guide essay
tips section. Many people come back with more detailed essays in future
cohorts that show they are strongly aligned with Outreachy's program
goals. It makes sense to encourage people right off the bat to do that.

Update tests to look for the new wording in the alignment rejection
ssection.