Allow viewing initial applications until interns start
When reviewing the time commitments of students who have been selected
as interns, it can be helpful to look at the time commitments stated by
their classmates. We do that by viewing the classmates' initial
applications. Allow the initial applications to be viewed until the date
the interns are ready to start.

Why cut off the view date at the date when interns start? Well, we have
to cut off the date somewhere in order to do this query. Otherwise the
query will return multiple RoundPage objects. Setting the end date to
when internship end date is too far out, because the initial application
period may be open while the previous cohort is still running.

If for some reason we need to view initial applications after the
interns start, we may be able to push this date out as late as 5 weeks
after the internship starts.