Add email reminder to mentors who haven't picked an intern
On the day of the intern selection deadline, send a reminder to all
mentors of projects that:

 - were approved to participate in this internship cohort
 - had at least one applicant fill out a final application
 - has at least one applicant who is unrated or rated 3 or above

Remind the mentors that the deadline to select an intern is today.
Encourage them to tell us if they aren't going to select an intern, or
if they're waiting on a team meeting, what date they expect to select an
intern by.

Also send the email to Outreachy organizers. This will help us keep
track of which mentors we need to manually pester.

This automated email is better than sending an email to all project
mentors, because mentors that have selected an intern may get panicked
if they see a second email about the intern selection deadline.

There could be the rare case we miss, where a mentor wanted to select
multiple interns for the same project. In that case, they wouldn't
receive this email. But this is the best we can do.