Update Marina's picture
Marina used this selfie for her bio in her All Things Open talk in
November 2021. Karen believes Marina liked this picture, as it
represented a picture of when she had cancer and still felt vibrant. The
past picture we had of Marina on the Outreachy website is over 15 years
old.