Allow each internship cohort to have a different sponsorship amount per intern
In prior internship cohorts, communities sponsored Outreachy interns at
$6,500 per intern. $6,000 went to the internship stipend, and $500 was
an administrative fee. That cost has remained the same since the first
Outreachy cohort in 2010 (back then the program was called "Outreach
Program for Women").

That means Outreachy has not given interns a stipend increase for over
10 years. With the cost of living rising around the world, we decided to
increase our intern stipend. In the December 2021 cohort, the internship
stipend was increased from $6,000 to $7,000.

A year later, we're increasing the sponsorship cost per intern and the
Outreachy sponsorship levels to reflect our increased internship
stipend.

Sponsorship per intern will now be $8,000 per intern. This covers the
$7,000 internship stipend and a $1,000 admin fee.

The admin fee increase reflects the increased time and resources
required to run the program. Outreachy organizers are spending on
promoting the program to potential applicants, working with mentors and
interns who have issues with their internships, running chats, and
generally working to support and improve the program in myriad ways.
Outreachy organizers have greatly increased our activities and promotion
of the program since 2010, which we believe has contributed to the high
quality and quantity of our applicants. The admin fee helps Software
Freedom Conservancy's administrative and legal support of Outreachy
which has also greatly increased as the program has grown.

In support of the sponsorship amount per intern increasing, the
following changes have been made to the Outreachy website:

 - Add a new "sponsorship_per_intern" field to the class that represents
   the internship cohort information (the RoundPage class)

 - Remove the hard-coded 6500 intern sponsorship amount from all code,
   and make that code reference the amount in the RoundPage object
   instead

 - Add tests to ensure:
   - the correct number of interns is shown on the project list page
   - the community read-only page shows the correct number of interns
     funded by the community sponsors
   - the staff pages for showing sponsorship amounts to invoice for is
     correct, both for older sponsorship pages with the old sponsorship
     per intern amount, and for newer sponsorship pages with the
     increased amount.
   - the staff dashboard section that shows community progress should
     show the correct number of interns the community has funding for
   - the community participation sign-up form should show the correct
     sponsorship amount per intern for that cohort
   - the community guide should mention the increased sponsorship cost
     for the most current cohort
   - the email to mentors and coordinators to announce the call for
     community participation should include the correct sponsorship cost
     per intern for the new cohort

 - I did not add tests to ensure the old blog posts show the correct
   cost-per intern. I will test those manually after I push the code.
   For completeness, here is the current set of old blog posts impacted:
     - 2021-01-15-cfp
     - 2021-08-18-cfp
     - 2022-01-01-cfp