Change sponsorship levels and sponsorship invoicing dates
Starting in the December 2022 cohort, Outreachy is changing our
sponsorship tiers. Update our sponsorship page to reflect that.

Change the dates when we start invoicing sponsors. We will begin to
invoice sponsors sooner (during the contribution period) in order to
avoid invoicing during Outreachy organizers' busy times (the intern
selection period and start of the internship).