Change feedback #4 to ask for a project report, rather than intern progress
When reporting to grant makers and sponsors the outcome of a particular
intern project, it helps to have a short paragraph about what internship
goals were completed. That saves us time from looking up the final
report blog post (if one was ever created).

Ask mentors who are requesting an internship extension to fill out the
standard internship progress questions we normally ask.

Update the feedback #4 form that Outreachy organizers see to include the
new instructions for the project report.