Keep closed projects in the same section as open projects
Applicants and mentors get confused when projects "disappear" from the
project page. What happens is the mentor marks the project as closed to
new contributors, and the project moves down to a separate section on
that page for closed projects.

In addition to being confusing, moving the projects to a new section
also changes the project number (e.g. "Debian project #2") displayed on
the project list and community landing page.

Keep both closed and open projects in the same sections. Add the phrase
"(Closed)" before the project title. In the project skills block and
in the project details block, make sure to use the same language about
what a closed project is. Explain when applicants should continue making
contributions, and when they shouldn't start new contributions.

Update tests to reflect new headings, since there's no longer an "Open
Projects" header (instead it's "Internship Projects").

As an added bonus, listing projects in the same sections removes some
duplicate code in the templates.

For now, keep the variables passed into the templates with the same
name. That means "ontime_projects" now contains both open and closed
projects. In the future, we might want to rename those variables, but I
don't trust myself to get it correctly without writing more tests first.