Update to Django 3.2 and update Python packages
I can't currently update to Django 4.0 because django-formtools 2.4
includes a bug that breaks our usage case of conditional form steps on
the Outreachy initial application:

https://github.com/outreachy/website/issues/537
https://github.com/jazzband/django-formtools/issues/220

I also ran into an odd issue with the home/test_eligibility.py unit
test. The work-around is documented in the code, although Jamey Sharp
and I still have no clue as to what the root cause is.

All the other unit tests use the home/scenarios.py code, and none of
them seem to have issues.

The unit tests are now more verbose, showing the permission errors that
are uncaught, and 404 not found messages. In those cases, we're actually
trying to test that errors are thrown when pages aren't accessed
properly.

I could add more code to catch those permission errors, but
I'm too tired after wrestling with updates. That's future work.

At least we're running a Django version that is receiving security
updates now! Django 3.2 is marked as being a long-term support release,
until early 2024:

https://www.djangoproject.com/download/

However, it will be really important to upgrade to 4.2 in mid-2023, so
that we can be on the next LTS release before 3.2 is deprecated in early
2024.