Handle error cases when coordinators use new community sign-up form
Community coordinators are often confused or overwhelmed by our
community CFP page. The behavior I would like to see is that they find
their community in the list of past participating communities, click on
that page, and then sign up.

Instead, coordinators get overwhelmed by scrolling through the list.
They end up at the bottom of the Community CFP page, where there is a
button for *new* communities to sign up.

Then the coordinator attempts to submit their past community as a new
community. That fails because the name is the same, and the slugify
call throws an error.

The new behavior of the new community sign up form is:
 - add a form to the participation rules page, asking for the new
   community name
 - use jquery to offer suggestions of old community names for that field
 - if the community name input into the form matches an old community,
   attempt to either sign that community up to participate, or (if the
   community is already participating) redirect to the page for the
   community call for mentors and projects

This new behavior requires us to check if the comrade is an approved
community coordinator. That means the participating rules page needs to
require people to be logged in, when it didn't before.

Update the tests to cover three cases:

 - new community sign-up needs to input a community name
 - old community sign-up using new community form redirects to the
   participation sign-up page if the old community is not participating
 - old community sign-up using new community form redirects to the
   community call for projects page if the old community is participating